requirements.txt
anyascii==0.3.2
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.9.3
certifi==2023.5.7
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.20
django-ckeditor==6.6.1
django-ckeditor-5==0.2.8
django-cors-headers==3.11.0
django-crispy-forms==2.0
django-filter==21.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
import-export==0.3.1
l18n==2021.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.2
tablib==3.5.0
telepath==0.3.1
toml==0.10.2
tzdata==2021.5
urllib3==2.0.3
wagtail==2.16.3
wagtail-ckeditor==1.2.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
